Fix static SAIL builds.
While SAIL could be built statically, the static library that was produced
was broken since commit ce0e3d2824936013083fdc8219a1f3916cc3ebbc
("BMP: Move BMP implementation into a common library to re-use it in ICO").
The sail-codecs library could not be linked against because it didn't contain
the contents of bmp-common, which is not exported separately. This was
because bmp-common was built as a static library while everything that linked
against it was a CMake object library. bmp-common was therefore not pulled
into the final codec library.

This commit simplifies and fixes sail-codecs-archive and extends it so that
common code shared between codecs (like bmp-common) can be pulled into the
combined sail-codecs library too.

Common libraries must become object libraries in static builds; furthermore,
they have to be marked with the SAIL_COMMON_OBJECT_LIBRARY property.